Fix casing by relocating
